0712 add Commit

+++ test3/WebContent/board/detail.jsp
@@ -9,6 +9,12 @@
 <title>/board/detail.jsp</title>
 </head>
 <body>
+<%  
+	String id=(String)session.getAttribute("id");
+	if(id==null){
+		id="";
+	}
+%>
 <%
 	// /board/detail.jsp?num=x
 	
@@ -39,15 +45,20 @@
 		<td><%=dto.getRegdate() %></td>
 	</tr>
 </table>
-<a href="updateform.jsp?num=<%=dto.getNum() %>">수정</a>
-<a href="javascript:deleteCheck()">삭제</a>
-<script>
+<% if(id.equals(dto.getWriter())) {%>
+	<a href="updateform.jsp?num=<%=dto.getNum() %>">수정</a>
+	<a href="javascript:deleteCheck()">삭제</a>
+<%} %>
+	<script>
 	function deleteCheck(){
 		var isDelete=confirm("글을 삭제 하시겠습니까?");
 		if(isDelete){
 			location.href="delete.jsp?num=<%=dto.getNum() %>";
 		}
 	}
+
+
+
 </script>
 </body>
 </html>
